requirements.txt
torch>=2.0
numpy>=1.24
gymnasium>=0.29
pufferlib>=0.7
pyyaml
tensorboard
onnx
